fix: commit util.
