+++ Cargo.toml
@@ -36,6 +36,8 @@ members = [
     "protocol-units/bridge/service",
     "protocol-units/bridge/grpc",
     "protocol-units/bridge/integration-tests",
+    "protocol-units/bridge/indexer-db",
+    "protocol-units/bridge/util",
     "protocol-units/settlement/mcr/runner",
     "benches/*",
 ]
@@ -53,11 +55,13 @@ rust-version = "1.75"
 [workspace.dependencies]
 # internal
 ## bridge
+bridge-util = { path = "protocol-units/bridge/util" }
 bridge-config = { path = "protocol-units/bridge/config" }
 bridge-service = { path = "protocol-units/bridge/service" }
 bridge-setup = { path = "protocol-units/bridge/setup" }
 bridge-integration-tests = { path = "protocol-units/bridge/integration-tests" }
 bridge-grpc = { path = "protocol-units/bridge/grpc" }
+bridge-indexer-db = { path = "protocol-units/bridge/indexer-db" }
 ## buildtime
 buildtime = { path = "util/buildtime" }
 buildtime-helpers = { path = "util/buildtime/buildtime-helpers" }
@@ -311,8 +315,9 @@ globset = "0.4.15"
 glob = "0.3.1"
 
 # trying to pin diesel
-# diesel = "=2.1.1"
-# migrations_internals = "=2.1.1"
+diesel = { version = "=2.1.1", features = ["postgres", "sqlite", "numeric", "r2d2"]}
+diesel_migrations = { version = "=2.1.0"}
+bigdecimal = "0.4.0"
 num_cpus = "=1.16.0"
 ahash = "=0.8.11"
