+++ protocol-units/bridge/service/src/actions.rs
@@ -1,66 +1,9 @@
-use crate::chains::bridge_contracts::{BridgeContract, BridgeContractError};
 use crate::chains::movement::utils as movement_utils;
-use crate::types::{Amount, BridgeAddress, BridgeTransferId, HashLock,HashLockPreImage};
-use crate::ChainId;
-use std::fmt;
+use bridge_util::actions::*;
+use bridge_util::chains::bridge_contracts::BridgeContract;
+use bridge_util::types::BridgeAddress;
 use std::future::Future;
 use std::pin::Pin;
-use thiserror::Error;
-
-#[derive(Error, Debug, Clone)]
-pub struct ActionExecError(TransferAction, BridgeContractError);
-
-impl ActionExecError {
-	pub fn inner(self) -> (TransferAction, BridgeContractError) {
-		(self.0, self.1)
-	}
-}
-
-impl fmt::Display for ActionExecError {
-	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		write!(f, "Action: {}/ Error: {}", self.0, self.1,)
-	}
-}
-
-#[derive(Debug, Clone)]
-pub struct TransferAction {
-	pub chain: ChainId,
-	pub transfer_id: BridgeTransferId,
-	pub kind: TransferActionType,
-}
-impl fmt::Display for TransferAction {
-	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		write!(f, "Action: {}/{}/{}", self.chain, self.transfer_id, self.kind)
-	}
-}
-
-#[derive(Debug, Clone)]
-pub enum TransferActionType {
-	LockBridgeTransfer {
-		bridge_transfer_id: BridgeTransferId,
-		hash_lock: HashLock,
-		initiator: BridgeAddress<Vec<u8>>,
-		recipient: BridgeAddress<Vec<u8>>,
-		amount: Amount,
-	},
-	WaitAndCompleteInitiator(u64, HashLockPreImage),
-	RefundInitiator,
-	TransferDone,
-	NoAction,
-}
-
-impl fmt::Display for TransferActionType {
-	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		let act = match self {
-			TransferActionType::LockBridgeTransfer { .. } => "LockBridgeTransfer",
-			TransferActionType::WaitAndCompleteInitiator(..) => "WaitAndCompleteInitiator",
-			TransferActionType::RefundInitiator => "RefundInitiator",
-			TransferActionType::TransferDone => "TransferDone",
-			TransferActionType::NoAction => "NoAction",
-		};
-		write!(f, "{}", act)
-	}
-}
 
 pub fn process_action<A>(
 	action: TransferAction,
@@ -79,10 +22,9 @@ where
 			amount,
 		} => {
 			let future = async move {
-
 				if recipient.0.len() == 32 {
 					if let Err(e) = movement_utils::fund_recipient(&recipient).await {
-					return Err(ActionExecError(action.clone(), e));
+						return Err(ActionExecError(action.clone(), e));
 					}
 				}
