+++ protocol-units/bridge/service/src/chains/ethereum/event_monitoring.rs
@@ -1,19 +1,19 @@
 use super::types::EthAddress;
-use crate::chains::bridge_contracts::BridgeContractError;
-use crate::chains::bridge_contracts::BridgeContractEvent;
-use crate::chains::bridge_contracts::BridgeContractMonitoring;
-use crate::chains::bridge_contracts::BridgeContractResult;
 use crate::chains::ethereum::types::AtomicBridgeCounterpartyMOVE;
 use crate::chains::ethereum::types::AtomicBridgeInitiatorMOVE;
-use crate::types::HashLockPreImage;
-use crate::types::LockDetails;
-use crate::types::{BridgeAddress, BridgeTransferDetails, BridgeTransferId, HashLock};
 use alloy::eips::BlockNumberOrTag;
 use alloy::primitives::Address;
 use alloy::providers::Provider;
 use alloy::providers::ProviderBuilder;
 use alloy_network::EthereumWallet;
 use bridge_config::common::eth::EthConfig;
+use bridge_util::chains::bridge_contracts::BridgeContractError;
+use bridge_util::chains::bridge_contracts::BridgeContractEvent;
+use bridge_util::chains::bridge_contracts::BridgeContractMonitoring;
+use bridge_util::chains::bridge_contracts::BridgeContractResult;
+use bridge_util::types::HashLockPreImage;
+use bridge_util::types::LockDetails;
+use bridge_util::types::{BridgeAddress, BridgeTransferDetails, BridgeTransferId, HashLock};
 use futures::SinkExt;
 use futures::{channel::mpsc::UnboundedReceiver, Stream, StreamExt};
 use std::{pin::Pin, task::Poll};
