+++ protocol-units/bridge/service/src/chains/mod.rs
@@ -1,3 +1,3 @@
-pub mod bridge_contracts;
 pub mod ethereum;
 pub mod movement;
+pub use bridge_util::chains::*;
