+++ protocol-units/bridge/service/src/chains/movement/event_monitoring.rs
@@ -1,20 +1,20 @@
 use super::client::MovementClient;
 use super::client_framework::FRAMEWORK_ADDRESS;
 use super::utils::MovementAddress;
-use crate::chains::bridge_contracts::BridgeContractError;
-use crate::chains::bridge_contracts::BridgeContractEvent;
-use crate::chains::bridge_contracts::BridgeContractEventType;
-use crate::chains::bridge_contracts::BridgeContractMonitoring;
-use crate::chains::bridge_contracts::BridgeContractResult;
-use crate::types::Amount;
-use crate::types::AssetType;
-use crate::types::BridgeAddress;
-use crate::types::BridgeTransferDetails;
-use crate::types::BridgeTransferId;
-use crate::types::HashLock;
-use crate::types::HashLockPreImage;
-use crate::types::LockDetails;
-use crate::types::TimeLock;
+use bridge_util::chains::bridge_contracts::BridgeContractError;
+use bridge_util::chains::bridge_contracts::BridgeContractEvent;
+use bridge_util::chains::bridge_contracts::BridgeContractEventType;
+use bridge_util::chains::bridge_contracts::BridgeContractMonitoring;
+use bridge_util::chains::bridge_contracts::BridgeContractResult;
+use bridge_util::types::Amount;
+use bridge_util::types::AssetType;
+use bridge_util::types::BridgeAddress;
+use bridge_util::types::BridgeTransferDetails;
+use bridge_util::types::BridgeTransferId;
+use bridge_util::types::HashLock;
+use bridge_util::types::HashLockPreImage;
+use bridge_util::types::LockDetails;
+use bridge_util::types::TimeLock;
 use anyhow::Result;
 use aptos_sdk::rest_client::aptos_api_types::VersionedEvent;
 use aptos_sdk::types::account_address::AccountAddress;
