+++ protocol-units/bridge/setup/Cargo.toml
@@ -12,6 +12,7 @@ rust-version.workspace = true
 
 [dependencies]
 bridge-config = { workspace = true }
+bridge-util = { workspace = true }
 bridge-service = { workspace = true }
 dot-movement = { workspace = true }
 commander = { workspace = true }
