+++ protocol-units/bridge/util/src/chains/mod.rs
@@ -1,0 +1,1 @@
+pub mod bridge_contracts;
